requirements.txt
pytest>=7.0.0
pytestomatio==2.10.1b1
selenium>=4.0.0
requests>=2.28.0
pytest-html>=3.1.0
python-dotenv>=0.19.0
allure-pytest>=2.10.0